multi-agent support added (#1)
* Refactor structure, add enum helpers, improve error handling, migrate user IDs to long

* First API works: /auth/register

* Enhance auth security with salted passwords and JWT config

* Add LLM module for language recognition and translation services with configuration and fallback support

* Improve API response handling with structured JSON parsing and case-insensitive deserialization

* Centralize case-insensitive JSON deserialization configuration across projects

* Update LLM configuration service naming and improve fallback model handling

* Enhance word explanation with structured data and Markdown support

* Remove redundant markdown explanation field to simplify data model

* Add documents to .gitignore and enhance Word entity for extended language support and explanations

* Enhance LLM explanation reliability with retries and add word data population endpoint

* multi-agent support added

+++ src/NewWords.Api/Repositories/IUserRepository.cs
@@ -7,7 +7,7 @@ namespace NewWords.Api.Repositories
 {
     public interface IUserRepository : IRepositoryBase<User>
     {
-        Task<User?> GetByIdAsync(int userId);
+        Task<User?> GetByIdAsync(long userId);
         Task<User?> GetByEmailAsync(string email);
         Task<PageData<User>> GetPagedUsersAsync(int pageSize, int pageNumber, bool isAsc = false);
     }
